move prop types to component files
